Remove app icon configuration from AppConstants.
* Moved AppIcon out of AppIconViewController
