+++ WordPress/Classes/ViewRelated/Me/App Settings/AppIconViewController.swift
@@ -170,66 +170,3 @@ open class AppIconViewController: UITableViewController {
         static let cellIdentifier = "IconCell"
     }
 }
-
-struct AppIcon {
-    let name: String
-    let isBordered: Bool
-    let isLegacy: Bool
-
-    var displayName: String {
-        return name.replacingMatches(of: " Classic", with: "")
-    }
-
-    var imageName: String {
-        let lowered = name.lowercased().replacingMatches(of: " ", with: "-")
-        return "\(lowered)-\(Constants.imageBaseName)"
-    }
-
-    static var isUsingCustomIcon: Bool {
-        return UIApplication.shared.alternateIconName != nil
-    }
-
-    static var currentOrDefault: AppIcon {
-        if let name = UIApplication.shared.alternateIconName {
-            return allIcons.first(where: { $0.name == name }) ?? defaultIcon
-        } else {
-            return defaultIcon
-        }
-    }
-
-    static var defaultIcon: AppIcon {
-        return AppIcon(name: AppIcon.defaultIconName,
-                       isBordered: false,
-                       isLegacy: false)
-    }
-
-    static var allIcons: [AppIcon] {
-        guard let bundleDict = Bundle.main.object(forInfoDictionaryKey: Constants.infoPlistBundleIconsKey) as? [String: Any],
-              let iconDict = bundleDict[Constants.infoPlistAlternateIconsKey] as? [String: Any] else {
-            return [defaultIcon]
-        }
-
-        let customIcons = iconDict.compactMap { (key, value) -> AppIcon? in
-            guard let value = value as? [String: Any] else {
-                return nil
-            }
-
-            let isBordered = value[Constants.infoPlistRequiresBorderKey] as? Bool == true
-            let isLegacy = value[Constants.infoPlistLegacyIconKey] as? Bool == true
-            return AppIcon(name: key, isBordered: isBordered, isLegacy: isLegacy)
-        }
-
-        return [defaultIcon] + customIcons
-    }
-
-    private enum Constants {
-        static let infoPlistBundleIconsKey = "CFBundleIcons"
-        static let infoPlistAlternateIconsKey = "CFBundleAlternateIcons"
-        static let infoPlistRequiresBorderKey = "WPRequiresBorder"
-        static let infoPlistLegacyIconKey = "WPLegacyIcon"
-        static let imageBaseName = "icon-app-60x60"
-    }
-
-    static let defaultIconName = "Blue"
-    static let defaultLegacyIconName = "WordPress"
-}
